+++ NEWS.md
@@ -1,5 +1,8 @@
 # cea 0.0.0.9000
 
+* Added `ci()` to calculate bootstrap confidence intervals from regression
+  model.
+  
 * Added `boot()` to calculate bootstrap replicates from regression model.
 
 * Added `INMB()` & `INHB()` to calculate INMB/INHB from regression model